# file-storager/docs/swagger.yaml
basePath: /
definitions:
  handlers.Analysis:
    properties:
      average_length_of_words:
        type: number
      average_sentences_per_paragraph:
        type: number
      average_words_per_sentence:
        type: number
      id:
        type: integer
      paragraphs_amount:
        type: integer
      sentences_amount:
        type: integer
      symbols_amount:
        type: integer
      words_amount:
        type: integer
    type: object
  handlers.FileExistsResponse:
    properties:
      exists:
        type: boolean
      id:
        type: integer
      status:
        type: string
    type: object
  handlers.FileStatusResponse:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: Service for managing stored files in DB.
  title: File Storage Manager
  version: "1.0"
paths:
  /files/{id}:
    get:
      description: Download file from DB.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
  /files/analysis:
    post:
      consumes:
      - application/json
      description: Save analysis result to DB.
      parameters:
      - description: Result of file analysis
        in: body
        name: analysis
        required: true
        schema:
          $ref: '#/definitions/handlers.Analysis'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
  /files/analysis/{id}:
    get:
      description: Get analysis result from DB. Result contains amount of paragraphs,
        sentences, words, symbols. Also contains average amount of sentences per paragraph,
        words per sentence, length of words.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.Analysis'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
  /files/exists/{id}:
    get:
      description: Check if file exists in DB.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FileExistsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileExistsResponse'
  /files/upload:
    post:
      description: Upload file to DB.
      parameters:
      - description: File to upload
        in: body
        name: file
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
  /files/wordcloud/{id}:
    get:
      description: Get word cloud from DB.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
    post:
      consumes:
      - image/png
      description: Save word cloud to DB.
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: integer
      - description: Word cloud to save
        in: formData
        name: wordCloud
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.FileStatusResponse'
swagger: "2.0"